fix type heigt variable

+++ src/main/java/org/openstreetmap/osmaxil/model/VegetationImport.java
@@ -98,11 +98,11 @@ public class VegetationImport extends AbstractImport {
 		this.leafType = leafType;
 	}
 
-	public Integer getHeight() {
+	public Float getHeight() {
 		return height;
 	}
 
-	public void setHeight(Integer height) {
+	public void setHeight(Float height) {
 		this.height = height;
 	}
